requirements.txt
beautifulsoup4==4.13.5
fastapi==0.116.1
huggingface_hub==0.34.4
ipython==8.12.3
numpy==1.26.4
playwright==1.54.0
pydantic==2.11.7
pymilvus==2.6.0
pymilvus_model==0.3.2
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.5
spacy==3.8.7
together==1.5.25
tqdm==4.67.1
uvicorn==0.35.0
